+++ backend/lib/openfisca/mapping/common.js
@@ -5,7 +5,7 @@ const { forEach } = require("../../../../data/all")
 const { generator } = require("../../../../lib/Dates")
 const {
   CONDITION_STATEGY,
-} = require("../../../../lib/Benefits/ComputeFrontEnd")
+} = require("../../../../lib/Benefits/ComputeJavascript")
 
 exports.isIndividuValid = function (individu, situation) {
   const age = moment(situation.dateDeValeur).diff(
