+++ lib/Benefits/Compute.js
@@ -7,7 +7,7 @@ const filter = require("lodash/filter")
 
 const moment = require("moment")
 const determineCustomizationIds = require("./Customization")
-const { computeJavascriptBenefits } = require("./ComputeFrontEnd")
+const { computeJavascriptBenefits } = require("./ComputeJavascript")
 /**
  * OpenFisca test cases separate ressources between two entities: individuals and families.
  * In Mes Aides, we don't care about this separation and want to show eligibilty results for the demandeur only.
