+++ lib/Benefits/ComputeJavascript.js
